Rename "policies" to "criteria" and handle them in a structured way.
"Criteria" is a clearer name for "what am I auditing for", and frees up
the term "policy" to be reused for the separate concept of "what is
enforced when I invoke `cargo vet`".

Moving criteria into the machine-readable domain unlocks a host of
benefits. In this commit, we introduce the ability to explicitly map
the criteria of imported audits to the local criteria. Later on, we'll
also use it to enforce different kinds of audits for different kinds of
crates.

This commit also introduces some sample policies.