Switch to toml_edit to get more control over serialization
This allows us to perform custom formatting behaviours in a visitor over
the serialized toml_edit::Document. Because of this extra control, we
can now make sure that dependency-criteria is rendered as an inline
table, and that strings are formatted as multiline strings when they
contain newlines, among other formatting changes we may want to make in
the future.

This patch does not use the format-preserving features of toml_edit, so
the output is always consistent for the same data.

Fixes #103
Fixes #155