Use GIT_EDITOR for editing and approving with cargo vet certify
This is an alternative approach to the `vet certify` UX which uses the
GIT_EDITOR in order to prompt and ask the user to input notes, and
accept a statement while also showing the criteria descriptions.

In order to function more reliably on Windows, this code attempts to
locate the git-for-windows install on the user's machine, and invokes
the user's GIT_EDITOR through the git-for-windows' shell environment.

This does not attempt to replace the interactive prompt for selecting
which criteria to certify for.

Fixes #165