Use the 'dirs' crate to store our cache in a platform-appropriate cache directory
This is a more persistent place to write the cache than in the temporary
directory, and we'll implement a GC mechanism on top of it to clean out
unnecessary cache files over time in a later commit.

This also renames some variables to reflect that the directory is now
explicitly a cache, rather a temporary directory.