Improve the output of guess-deeper
This makes a few changes to resolution to track our confidence in the
requirements for a specific dependency, and propagate that through to
emitting suggestions and vet error reports.

Lines corresponding to uncertain requirements will be dimmed when
written to a terminal, and will be ordered after confident lines when
listing.

This mode will be made the default in a follow-up.