More aggressively make parallelizable io async
This makes all io around networking and diffing crates async, and attempts to
run things concurrently as much as possible. For example, when generating
diffstats for suggest, all crate downloads and their diffs are run concurrently
with limits to avoid too many connections to crates.io at once.

This refactor also includes some extra logic which was added to the package
fetching logic to ensure that we update the mtime on the .crate file whenever
it is accessed, so that information can be used in the next patch for GC.

To make this change, the cache needed to be made shareable across multiple
concurrent tasks. The core mutable state was moved behind a mutex, and new
tables of async OnceCells were added to ensure that only one task would attempt
to download or diffstat each crate at a time.

With these changes, the time to perform a `cargo vet suggest` on
mozilla-central with an empty cache and use of the cargo registry cache
disabled, reduced from ~18s to ~1.5s.