Fix: Auditor information is not read from global git config
The git configuration is used (in `get_user_info()` in src/main.rs) by
calling `git config --get`, which is not the same as reading the global
git configuration, as `git config` considers the repository-local
configuration file as well as the $HOME git configuration file.

Hence, adapt the wording here.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>