Improve criteria selection in certify
This patch makes 2 major changes to certify:

First, certify will no longer attempt to guess the version you are
trying to certify unless there is a previous diff or inspect command in
history for the crate being certified. This is to avoid potential
mistakes, as it would be easy to miss the fact that the wrong version
was selected.

Second, when suggesting criteria to certify, the command will try
harder to find the relevant set of criteria. If the diff or version
being certified isn't found in the `suggest` output history, it will now
also check the unaudited table, and fall back to running the resolver to
determine what criteria would heal the graph when specified for this
revision.

This pass with the resolver is intended to be more relaxed that
`suggest`, as the precise delta being checked is already known, so it
should be able to give suggested criteria even for audits which would
not be suggested by the `suggest` algorithm.