Introduce an Out trait which abstracts over the program's output
This is certainly not a final solution. It doesn't integrate at all with our error reporting or tracing codepaths, and has the limitation that it requires mutable access thanks to using `io::Write` directly.

This allows us to mock out more of terminal interaction to allow it to be tested from within our unit tests.