Don't include prompt in lines read from terminal.
Don't call `console::read_line_initial_text` with the prompt as the
'initial text', as that causes it to include the prompt in the input:
entering "2" and hitting ENTER returns the string "> 2". Instead, just
write the prompt to the console, and then call `read_line`.

Rename `read_line_initial_text` to `read_line_with_prompt`.