Interactively display criteria and instructions for inspect and diff
This changes the behaviour of the inspect and diff instructions to now
predict the required criteria for the audit, and display the relevant
information up-front. A warning is shown if no criteria could be
determined, to help avoid accidentally unnecessary audits.

In addition, a partial audit for `syn` was added to the test project to
make it easier to test delta audit suggestions.

Fixes #200