Report JSON errors uniformly to Out
This changes the way that errors are handled and allows them to be
uniformly presented to the standard output when they're encountered.

This will also include panic errors, which should now also be reported
as error objects to the normal Output when encountered, and are
formatted using miette to fit in with other errors.

This is done by introducing a new hook which is used by cargo-vet to
report errors to the console, and handles overriding the format for json
errors. Errors printed to enter into things like logs will be presented
as normal, as logs are currently always destined for human processing.

In order to allow the `Out` object to be shared with the error reporter,
this ended up requiring the type to be wrapped in an `Arc<T>`, and for
the signatures of various methods to be changed. As the `io::Write`
base trait couldn't be used anymore, the signature of `write_fmt` was
also simplified to no longer return an error, simplifying the call-sites
of `writeln!(out` across the codebase.