Use a more transparent type for source files
This patch introduces our own implementation of miette's SourceCode
trait which allows accessing the string value and holds the name. This
is then used instead of NamedSource to allow things like parsers to run
on an `Arc<SourceFile>`, reducing string copies.

In addition, this SourceFile kind has support for doing nested files,
which will be used in a later commit to avoid unnecessary string copies
when reading files from the imports.lock file.