Store verbatim source in imports.lock
This patch changes how imports.lock is handled internally, such that
verbatim imported sources are now generally preserved for error
reporting purposes, and information like criteria descriptions are
consistently cached locally.

With this change, the imports.lock is now a list of URIs which have been
fetched and imported into this repository's supply-chain. This is used
for both imported criteria descriptions and imported audits. When
loading the store, imports are now resolved, being injected into the
relevant places where they were mentioned in the `ConfigFile` and
`AuditsFile` data structures so that other code can use then easily.

The handling of criteria descriptions in this way was done to avoid
needing to modify the imported source text, so that errors can continue
to be reported relative to the original imported source text.

Each imported URL starts with a block of TOML "front-matter" (delimited
by --- lines) which describes the URL and the number of lines. This was
used rather than a more concise format to allow recording cache keys,
metadata, or other sources in the future. The 'hrx' format was
considered for this file, however was not used, as the design is meant
to model archives and filesystem paths, meaning that URIs are not legal
file names (due to containing a ':' character).

After this change, the first run of `cargo vet` on your repository will
report a warning, and your `imports.lock` file will need to be
re-generated. This should happen automatically.