Improve the precision and exemption re-use of "regenerate exemptions"
This patch reworks the `minimize_exemptions` logic to be more precise
and fix issues such as those found in #292.

This new algorithm doesn't use the `suggest` backend, instead
iteratively selecting exemptions which are required to allow resolve to
proceed, and then re-trying the resolver as-necessary. This is an
improvement over the existing logic as it is able to be much more
precise while prioritizing keeping exemptions pinned in the past over
crate sizes.

The algorithm also has logic to attempt to preserve some special
exemptions, such as those with `dependency_criteria` or which have been
marked `suggest = false`, so that they are not lost when regenerating,
while not allowing the criteria they're exempting to expand (as that
would be potentially surprising for both cases).

If there are no existing audits for a crate which can be used as a
guideline for a newly-required audit, the new algorithm will always bias
toward doing a full-exemption for the current crate version, over trying
to pick an arbitrary node to audit, as this simplifies the logic
substantially and should produce acceptable results for a newly added
dependency subtree.

Unlike the previous approach, there should be no cases where `vet` fails
after a successful `cargo vet regenerate`, as it correctly handles
`dependency_criteria` due to the iterative resolver calls.

If we decide to eventually run this automatically when a `vet` is
successful (e.g. in `cargo vet certify` to fix #262), it would be fairly
straightforward to extend this with an option to prevent the possibility
of new exemptions being added.

Fixes #292