Stop using 0.0.0 as the root version
The logic in the resolver previously treated the version 0.0.0 as the
"root version" (i.e. equivalent to no version at all). Unfortunatley,
this version can be published, meaning that we'll misbehave when trying
to interact with an actually published version 0.0.0 of a crate, and
will always consider it fully-audited as the "root version".

This patch updates the logic in the resolver to operate on
`Option<Version>` instead of `Version` in places where the root version
is valid.

Because this encoding of 0.0.0 as the root version is saved on disk in
the diff-cache, this patch changes the diff-cache format in an
incompatible way to force it to be re-built. It also adds a mechanism to
make forcing the diff-cache to be re-built easier in the future.

Unfortunately, these changes lead to a large number of JSON output test
failures, as we exposed the old internal desugaring of '0.0.0' as a diff
source in the JSON output, which has now changed to null. We may want to
look into changing this part of the output to avoid talking about full
audits as "diffs from null" in the future.