Implement cargo vet aggregate
This feature allows specifying multiple sources in a `sources.list` file
which will be processed to generate a new source file, containing all of
the audits from each listed source. Criteria are merged from these
sources by-name, with the descriptions and implies sets requried to be
identical between criteria with the same name from different sources.

The chain of what imports were used for a given criteria or audit is now
tracked in the .toml file as well, using an aggregated-from field on the
relevant entries.

Fixes #83