Relax parsing when importing remote audits
This will allow the audits files to still partially parse and provide
audits when the cargo-vet version used upstream is newer than the local
version in more cases. The goal is for this to avoid issues with
organizations updating the version of cargo-vet they are using
out-of-sync with one another.