Filter diffs to allow ignoring certain generated files
This will be more important once we support git revisions, as we won't
be able to generate these files for git revisions, leading to
unnecessarily large diffs.

As we want to re-generate diffstats to use the new format, this also
updates the diffstat version number, and takes the opportunity to change
the format to a more flexible one. This unfortunately leads to large
test output changes, including JSON output, making it a breaking change.