Diff each file separately, combining outputs with a pager
This greatly improves on the filtering capabilities which I built
in #371, making them much more reliable. Without this change, a poor
interaction between between orderfiles and --skip-to for diffs with
--no-index in git diff would cause some files to occasionally not be
ignored when they were supposed to be (such as Cargo.lock files) which
could lead to issues when used with some git revisions.

The new approach means that we always generate diffs for only the files
we intend to, in a more reliable manner. Like the GIT_EDITOR setup, this
required some work to integrate with the git-provided pager, however
most of the work could be re-used from GIT_EDITOR.

This will also make the `cargo vet diff` command interact better with
the --out-file command line option, as previously the output would
always be written to stdout.