Treat git revisions as distinct versions from published crates
When depending on a git revision, you are depending on a different
version of the crate than the one provided on crates.io. This patch adds
support for adding audits for specific git revisions which may exist in
your tree in order to provide actual audits for these versions.

This requires some integration into the diff filtering logic added in
previous revisions in order to make the experience somewhat ergonomic
and to not produce misleading diffs.