Replace dependency-criteria on audits with policy criteria
This patch removes dependency-criteria from the UI as something which
can be added to audits and exemptions. Instead, we now support adding
dependency-criteria to policy entries for third-party crates.

This will enable large simplifications to the core resolver in the
future, however those have not been implemented in this patch, nor have
potential improvements around the reliability of suggestions. This patch
just support for the feature, and adds support for policy entries.

Before the next release, we should require that policy entries for
third-party crates be versioned, however that has not yet been
implemented in this patch.