Remove the --shallow CLI option from cargo-vet
With dependency-criteria now removed, we should be able to accurately
predict the required criteria for every dependency in the dependency
graph, removing the need for uncertainty or the --shallow option.

This patch doesn't implement accurate guessing, only removing the
shallow option and updating test snapshots to reflect the new expected
behaviour.