Remove uncertainty and the blaming step from the resolver
This patch does some fairly-major changes to the underlying resolver
algorithm. Thanks to the removal of dependency-criteria from audits and
exemptions, it's now possible to pre-compute the required criteria for
every crate in the dependency graph ahead-of-time, and then validate
each crate's criteria separately.

This removes the need for the blame phase, as we already know the
responsible packages ahead of time, allowing these phases to be
disentangled from one-another and simplified.

As there is no more uncertainty, the CriteriaFailureSet type is no
longer necessary, and the output of a significant number of commands has
been re-ordered to reflect that.

In addition, the algorithm page in the book has been updated to reflect
the new implementation.