Only import the minimal number of audits required for vet to pass
This is a large change to how the import backend works. It takes
advantage of the lack of dependency-criteria to dramatically simplify
the set of audits which we import such that we always only import the
smallest number of audits possible.

While this impacts all imports, this is part of the work towards
wildcard audits, as reifying all wildcard audits would be unnecessarily
noisy, due to the number of releases of some crates.

In order to do this in a reliable fashion, we now sort audit edges more
aggressively when traversing the tree, so as to give a more
deterministic total order and ensure we're reliably selecting the same
edges.