Implement Wildcard Audits
This builds upon the changes in #399 to implement basic support for
wildcard audits. These can be specified in your audits.toml file under
the `wildcard-audits` key (rather than `audits`). A separate key is used
to ensure that wildcard audits are placed above normal audits in
audits.toml, to make them more discoverable.

Wildcard audits take a crates.io user-id and date range, and act as an
audit for any crates published by the user with that user-id in that
date range.

In order to avoid hitting the crates.io API too often, we only fetch
publisher information about crates when needed, and maintain a local
cache of publisher information in the cache directory.

While all information about publishers for each version is stored in the
cache directory, only publisher information which is necessary will be
stored in the imports.lock file for use in --locked mode. This is done
in a similar manner to other imported audits since #399 by tracking
which audit edges are used when running the resolver.

Wildcard audit end dates are allowed to be specified at most 1 year in
the future.