Add an option to create wildcard audits with certify
This adds a --wildcard flag to the certify subcommand which can be used
to create wildcard audits. Start and end dates must be specified with
--start-date and --end-date, though they have fallback values.

This is a first-draft of the UX, and we may improve it in the future.
We expect creating wildcard audits to be less frequent than other types,
so the experience is a bit less polished.

Fixes #402