Update installation instructions to pass --locked to cargo install
Without this, installation will pick up any new semantically-compatible dependencies available at the time of compilation. We haven't been entirely rigorous with audits for the tool itself, but we should in principle aim to move in that direction, and therefore should be encouraging people to build with the same Cargo.lock as us.