Avoid unrelated imports during certify and vet
This is done by making whether imports or exemptions are preferred
configurable on a per-package basis when running regenerate_exemptions.

This combines the existing iterative approach, which prefers imports,
with an approach based on used edges when auditing, which prefers local
audits. After the change, it is still possible for `vet` to cause
imports to occur if they are necessary to vet successfully, however it
will not cause unrelated exemptions to be removed.

Fixes #391