Pass a high visibility parameter when loading sourcegraph diffs.
Sourcegraph currently divides segments long diffs into multiple pages,
with a "show more" button at the bottom (see [1] for an example).

Unfortunately, the "show more" button is small and easy to miss, which
could result in people mistakenly auditing only a subset of the diff. As
an interim fix, Sourcegraph just recommends passing a high value for
the "visibility" parameter. This parameter controls the number of files
whose diff is displayed, and assuming it's higher than the number of
actual files present in the diff, will redirect to the latter (so
passing ?visiblity=1000000 will result in ?visibility=24 showing in the
URL bar in the below example).

[1] https://sourcegraph.com/crates/async-task/-/compare/v4.0.3...v4.3.0