Emit an error when trying to remap remote built-in criteria
This was never properly implemented, and is being changed to emit a
useful error, rather than producing confusing behaviour, such as
accidentally creating an implies cycle between the safe-to-run and
safe-to-deploy criteria.

This patch also removes documentation suggesting support for multiple
criteria in criteria-map entries. This has never been implemented, and
would not be trivial to implement with the current infrastructure, so
has been removed. Some internal code has also been simplified to produce
a better error in that situation.