Map criteria into the local namespace eagerly when importing
This removes the need for criteria namespaces entirely, and also allows us to
implement mapping behaviour over the built-in criteria without bloating the
criteria set.

Unfortunately, this causes changes in the (quite-unstable) violation JSON
output. We'll probably want to change it in the future to not partially use
internal data structures so that it can be more stable.

Fixes #425
Fixes #430