Add support for registry suggestions
This adds suggestions from the registry to the `cargo vet` and `cargo vet
suggest` suggestion output. This is done by fetching the registry and the
corresponding audits files when generating suggestions, and comparing the
suggestion output from before and after to see if it would make an improvement.

Currently this adds no information to the JSON output to encode the
registry suggestion information, though we may want to encode it somehow
in the future.

Fixes #104