Track the version of cargo-vet used to create the supply-chain store
This is important given changes like #430, which will change the meaning of
imports.lock in such a way that earlier versions of cargo-vet may not
understand parts of it correctly anymore. While we generally need to keep the
audits.toml file somewhat forward-and-back compatible due to imports, this is
less of a concern for other files, so preventing downgrades could be valuable
here.

This new approach allows implicit upgrades when running without --locked, but
prevents downgrades or locked upgrades. cargo-vet versions with the same minor
release are considered compatible, so only the major and minor components are
considered.

Fixes #430