Default audit-as-crates-io to true on init if description or repository matches
This adds an extra layer of checking above name/version against
crates.io, of also checking if the crate's description or repository
appears to match.  This is done by querying the crates.io API to get the
published description and repository for the most recent version.

This also required changing how we initialize exemptions when initialize
a new repo. The new logic now relies on the regenerate-exemptions logic,
and is run after fix_audit_as. This also required changing how test
setup works under the hood to create a mock store. We may want to tweak
this in the future.