Remove dynamic `implies` handling from criteria picker.
The code I'm removing does two things.

First, it highlights criteria implied by the selection is yellow. This
is nice in theory, but in practice most users just have safe-to-run and
safe-to-deploy, and so one option is always green and one option is
always yellow, which doesn't really convey the intended nuance.

Second, it removes implied criteria from the list of "chosen criteria".
Unfortunately, this is currently out of sync with the selector (which
lets you select both safe-to-deploy and safe-to-run and turn them both
green). We could fix this (and have the selector refuse to select
criteria that are implied by an already-selected criteria), but the
naive way to do this would lead to the user getting stuck when they want
to switch to safe-to-run. In general, it's easier to just be less fancy
on the input side and then collapse `implies` cases when generating the
entry (which the code already does).