Add --cargo-arg option for arguments to pass through to cargo
Some projects may require additional arguments in the cargo
metadata command in order to parse their Cargo.toml. For
example Chromium requires `-Zbindeps` as it uses the feature
for a dependency on cxxbridge-cmd. This allows such projects
to use `cargo vet --cargo-arg=<STUFF>` in order to add one or
more arguments to the cargo metadata command line.