Perform import validation when aggregating remote audit files
This copies the behaviour from normal imports, meaning that it handles
things like version mismatches gracefully, and only emits a warning that
those audits were ignored.

We may want to include an option to make this more aggressive, as
aggregate is expected to be frequently run in automation so warnings may
be invisible.