Include criteria which imply the required criteria in suggest output
For most consumers using exclusively the default criteria, this will
only impact the output for `safe-to-run` recommendations to mention that
`safe-to-deploy` would also be applicable. For consumers using more
complex criteria, this may increase the size of the suggest output more
substantially.

This change does impact JSON output, but only for strings generally
intended for human consumption. If a script was previously parsing these
strings it may be broken by the change.

Fixes #613