Remember user --mode choices, and support diff.rs for inspect
This change adds support to remember user's --mode choices, so that they
do not need to be used every time. In addition, it adds a basic support
for diff.rs to inspect (by loading a diff from a version to itself) such
that this mode memory can be shared between diff and inspect.

If xfbs/diff.rs#24 is merged, we should change the diff.rs inspect
behaviour to use the new browse endpoints instead.

Finally, this patch changes the default mode from sourcegraph to
diff.rs. As noted in #611, sourcegraph has been quite unreliable lately,
and diff.rs is likely to provide a better experience until that is
resolved.