Trusted Publisher support for wildcard audits and trusted entries
This is done by replacing the previous user ID types with a new type
which also can hold a string to identify a wildcard/trusted publisher.
This validation is based on the 'trustpub_data' field exposed in the
crates.io API.

A simple string is used for ergonomics to represent a trusted publisher,
however the specific representation is TBD for each future provider, to
align better with the data structure used internally by cargo and
crates.io.

Fixes #663